requirements.txt
python-dotenv
pymongo
tqdm
colorama